--- gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sun Nov 26 17:20:21 CET 2017
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-3.4-bin.zip
