--- modules/core/db/update/mysql/17/171126-0-dropHrtrainingplatformQuestionCategory.sql
@@ -1,1 +1,0 @@
-drop table if exists HRTRAININGPLATFORM_QUESTION_CATEGORY cascade ;
