--- modules/core/db/update/mysql/17/171126-0-dropHrtrainingplatformSelectedAnswer.sql
@@ -1,1 +1,0 @@
-drop table if exists HRTRAININGPLATFORM_SELECTED_ANSWER cascade ;
