--- modules/core/db/update/mysql/17/171126-0-dropHrtrainingplatformSurveyQuestionLink.sql
@@ -1,1 +1,0 @@
-drop table if exists HRTRAININGPLATFORM_SURVEY_QUESTION_LINK cascade ;
