--- modules/global/src/com/company/hrtrainingplatform/entity/QuestionCategory.java
@@ -1,26 +1,0 @@
-package com.company.hrtrainingplatform.entity;
-import javax.persistence.Entity;
-import javax.persistence.Table;
-import javax.persistence.Column;
-import com.haulmont.cuba.core.entity.StandardEntity;
-import com.haulmont.chile.core.annotations.NamePattern;
-
-@NamePattern("%s|name")
-@Table(name = "HRTRAININGPLATFORM_QUESTION_CATEGORY")
-@Entity(name = "hrtrainingplatform$QuestionCategory")
-public class QuestionCategory extends StandardEntity {
-    private static final long serialVersionUID = 7404944882434777834L;
-
-    @Column(name = "NAME")
-    protected String name;
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-
-}
