--- modules/global/src/com/company/hrtrainingplatform/entity/Test.java
@@ -1,27 +1,0 @@
-package com.company.hrtrainingplatform.entity;
-
-import javax.persistence.Entity;
-import javax.persistence.Table;
-import javax.persistence.Column;
-import com.haulmont.cuba.core.entity.StandardEntity;
-import com.haulmont.chile.core.annotations.NamePattern;
-
-@NamePattern("%s|name")
-@Table(name = "HRTRAININGPLATFORM_TEST")
-@Entity(name = "hrtrainingplatform$Test")
-public class Test extends StandardEntity {
-    private static final long serialVersionUID = 7966127297266077716L;
-
-    @Column(name = "NAME", nullable = false)
-    protected String name;
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-
-}
