+++ modules/global/src/com/company/hrtrainingplatform/entity/Training.java
@@ -50,6 +50,9 @@ public class Training extends StandardEntity {
     @Column(name = "STATUS", nullable = false)
     protected String status;
 
+    @OneToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "SURVEY_ID")
+    protected Survey survey;
 
     public void setDescription(String description) {
         this.description = description;
@@ -100,7 +103,13 @@ public class Training extends StandardEntity {
         return status == null ? null : STATUS.fromId(status);
     }
 
+    public void setSurvey(Survey survey) {
+        this.survey = survey;
+    }
 
+    public Survey getSurvey() {
+        return survey;
+    }
 
 
 }
