+++ modules/global/src/com/company/hrtrainingplatform/entity/messages.properties
@@ -16,6 +16,7 @@ HREmployee = H R Employee
 Employee.user = User
 Employee = Employee
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
 Survey = Survey
 Book.cost = Cost
@@ -26,6 +27,9 @@ Book.title = Title
 Book.publisher = Publisher
 Book.year = Year
 >>>>>>> parent of 957ec06... Merge branch 'master' into Amir
+=======
+Survey = Survey
+>>>>>>> parent of ba4f4c5... survey class by Amir + Ricardo
 Certificate.description = Description
 Certificate = Certificate
 Certificate.file = File
